refactor(blockifier): global execute_next_deprecated_syscall (#6023)
